confirm delete actions

+++ application/views/calendar.php
@@ -43,7 +43,7 @@ echo ($this->input->get('me') == 1) ? '&amp;me=1' : '';
           </a>
         <?php endif; ?>
 
-        <a class="btn btn-danger" href="/leave/delete/<?php echo $c->id; ?>">
+        <a data-confirm-delete-url class="btn btn-danger" href="/leave/delete/<?php echo $c->id; ?>">
           <i class="far fa-trash-alt"></i>
         </a>
       </td>
@@ -83,7 +83,7 @@ echo ($this->input->get('me') == 1) ? '&amp;me=1' : '';
           </a>
         <?php endif; ?>
 
-        <a class="btn btn-danger" href="/transports/delete/<?php echo $c->id; ?>">
+        <a data-confirm-delete-url class="btn btn-danger" href="/transports/delete/<?php echo $c->id; ?>">
           <i class="far fa-trash-alt"></i>
         </a>
       </td>
