+++ application/views/contacts/show.php
@@ -9,7 +9,7 @@ ob_start();
     <a class="btn btn-outline-primary" href="/contact/edit/<?php echo $contact->id; ?>" role="button">Modifier</a>
   <?php endif; ?>
   <?php if ($controller->hasPermission('contacts', 'delete')): ?>
-    <a class="btn btn-outline-danger" href="/contact/delete/<?php echo $contact->id; ?>" role="button">Supprimer</a>
+    <a data-confirm-delete-url class="btn btn-outline-danger" href="/contact/delete/<?php echo $contact->id; ?>" role="button">Supprimer</a>
   <?php endif; ?>
 </h1>
 <p class="lead">Affichage des informations à propos de ce contact</p>
