+++ application/views/transports/list.php
@@ -39,7 +39,7 @@ ob_start();
           </a>
         <?php endif; ?>
 
-        <a class="btn btn-danger" href="/transports/delete/<?php echo $c->id; ?>">
+        <a data-confirm-delete-url class="btn btn-danger" href="/transports/delete/<?php echo $c->id; ?>">
           <i class="far fa-trash-alt"></i>
         </a>
       </td>
